+++ lib/src/controller/map_controller.dart
@@ -2,32 +2,36 @@ import 'package:flutter/material.dart';
 import 'package:flutter_osm_plugin/flutter_osm_plugin.dart';
 import 'package:flutter_osm_plugin/src/controller/osm_controller.dart';
 
+import 'base_map_controller.dart';
+
+
 /// class [MapController] : map controller that will control map by select position,enable current location,
 /// draw road , show static geoPoint,
 ///
 /// [initMapWithUserPosition] : (bool) if is true, map will show your current location
 ///
 /// [initPosition] : (GeoPoint) if it isn't null, the map will be pointed at this position
-class MapController {
-  late OSMController _osmController;
-  final bool initMapWithUserPosition;
-  final GeoPoint? initPosition;
-
+class MapController extends BaseMapController {
   MapController({
-    this.initMapWithUserPosition = true,
-    this.initPosition,
-  }) : assert(initMapWithUserPosition || initPosition != null);
+    bool initMapWithUserPosition = true,
+    GeoPoint?initPosition,
+  })  : assert(
+          initMapWithUserPosition || initPosition != null,
+        ),
+        super(
+          initMapWithUserPosition: initMapWithUserPosition,
+          initPosition: initPosition,
+        );
 
+  @override
   void init(
-    OSMController osmController, {
-    initMapWithCurrentPosition = false,
-    initPosition,
-  }) {
-    _osmController = osmController;
+    OSMController osmController,
+  ) {
+    super.init(osmController);
   }
 
   void dispose() {
-    _osmController.dispose();
+    osmController.dispose();
   }
 
   /// initialise or change of position with creating marker in that specific position
@@ -35,77 +39,77 @@ class MapController {
   /// [p] : geoPoint
   ///
   Future<void> changeLocation(GeoPoint p) async {
-    await _osmController.changeLocation(p);
+    await osmController.changeLocation(p);
   }
 
   ///animate  to specific position with out add marker into the map
   ///
   /// [p] : (GeoPoint) position that will be go to map
   Future<void> goToLocation(GeoPoint p) async {
-    await _osmController.goToPosition(p);
+    await osmController.goToPosition(p);
   }
 
   ///remove marker from map of position
   /// [p] : geoPoint
   Future<void> removeMarker(GeoPoint p) async {
-    _osmController.removeMarker(p);
+    osmController.removeMarker(p);
   }
 
   ///change Icon Marker
   /// we need to global key to recuperate widget from tree element
   /// [key] : (GlobalKey) key of widget that represent the new marker
   Future changeIconMarker(GlobalKey key) async {
-    await _osmController.changeIconMarker(key);
+    await osmController.changeIconMarker(key);
   }
 
   /// change static position in runtime
   ///  [geoPoints] : list of static geoPoint
   ///  [id] : String of that list of static geoPoint
   Future<void> setStaticPosition(List<GeoPoint> geoPoints, String id) async {
-    await _osmController.setStaticPosition(geoPoints, id);
+    await osmController.setStaticPosition(geoPoints, id);
   }
 
   /// zoom in/out
   /// [zoom] : (double) positive value:zoomIN or negative value:zoomOut
   Future<void> zoom(double zoom) async {
-    await _osmController.zoom(zoom);
+    await osmController.zoom(zoom);
   }
 
   /// zoomIn use defaultZoom
   /// positive value:zoomIN
   Future<void> zoomIn() async {
-    await _osmController.zoomIn();
+    await osmController.zoomIn();
   }
 
   /// zoomOut use defaultZoom
   /// negative value:zoomOut
   Future<void> zoomOut() async {
-    await _osmController.zoom(-1);
+    await osmController.zoom(-1);
   }
 
   /// activate current location position
   Future<void> currentLocation() async {
-    await _osmController.currentLocation();
+    await osmController.currentLocation();
   }
 
   /// recuperation of user current position
   Future<GeoPoint> myLocation() async {
-    return await _osmController.myLocation();
+    return await osmController.myLocation();
   }
 
   /// enabled tracking user location
   Future<void> enableTracking() async {
-    await _osmController.enableTracking();
+    await osmController.enableTracking();
   }
 
   /// disabled tracking user location
   Future<void> disabledTracking() async {
-    await _osmController.disabledTracking();
+    await osmController.disabledTracking();
   }
 
   /// pick Position in map
   Future<GeoPoint> selectPosition() async {
-    GeoPoint p = await _osmController.selectPosition();
+    GeoPoint p = await osmController.selectPosition();
     return p;
   }
 
@@ -123,7 +127,7 @@ class MapController {
     Color? roadColor,
     double? roadWidth,
   }) async {
-    return await _osmController.drawRoad(
+    return await osmController.drawRoad(
       start,
       end,
       roadColor: roadColor,
@@ -133,69 +137,69 @@ class MapController {
 
   ///delete last road draw in the map
   Future<void> removeLastRoad() async {
-    await _osmController.removeLastRoad();
+    await osmController.removeLastRoad();
   }
 
   /// draw circle into map
   Future<void> drawCircle(CircleOSM circleOSM) async {
-    await _osmController.drawCircle(circleOSM);
+    await osmController.drawCircle(circleOSM);
   }
 
   /// remove specific circle in the map
   Future<void> removeCircle(String keyCircle) async {
-    await _osmController.removeCircle(keyCircle);
+    await osmController.removeCircle(keyCircle);
   }
 
   /// draw rect into map
   Future<void> drawRect(RectOSM rectOSM) async {
-    await _osmController.drawRect(rectOSM);
+    await osmController.drawRect(rectOSM);
   }
 
   /// remove specific region in the map
   Future<void> removeRect(String keyRect) async {
-    await _osmController.removeRect(keyRect);
+    await osmController.removeRect(keyRect);
   }
 
   /// remove all rect shape from map
   Future<void> removeAllRect() async {
-    return await _osmController.removeAllRect();
+    return await osmController.removeAllRect();
   }
 
   /// clear all circle
   Future<void> removeAllCircle() async {
-    await _osmController.removeAllCircle();
+    await osmController.removeAllCircle();
   }
 
   /// remove all shape from map
   Future<void> removeAllShapes() async {
-    await _osmController.removeAllShapes();
+    await osmController.removeAllShapes();
   }
 
   Future<void> advancedPositionPicker() async {
-    await _osmController.advancedPositionPicker();
+    await osmController.advancedPositionPicker();
   }
 
   /// select current position and finish advanced picker
   Future<GeoPoint> selectAdvancedPositionPicker() async {
-    return await _osmController.selectAdvancedPositionPicker();
+    return await osmController.selectAdvancedPositionPicker();
   }
 
   /// get current position
   Future<GeoPoint> getCurrentPositionAdvancedPositionPicker() async {
-    return await _osmController.getCurrentPositionAdvancedPositionPicker();
+    return await osmController.getCurrentPositionAdvancedPositionPicker();
   }
 
   /// cancel advanced picker
   Future<void> cancelAdvancedPositionPicker() async {
-    return await _osmController.cancelAdvancedPositionPicker();
+    return await osmController.cancelAdvancedPositionPicker();
   }
-  /*
+/*
   /// draw road manually
   ///  [path] : (list) path of the road
   Future<void> drawRoadManually(
     List<GeoPoint> path,
   ) async {
     assert(path.length > 3);
-    await _osmController.drawRoadManually(path);
+    await osmController.drawRoadManually(path);
   }*/
 }